Move :codeviewer into sdk, introduce :hightlights-core
