--- compose/codeviewer/api/codeviewer.klib.api
@@ -1,13 +1,0 @@
-// Klib ABI Dump
-// Targets: [wasmJs]
-// Rendering settings:
-// - Signature version: 2
-// - Show manifest properties: true
-// - Show declarations: true
-
-// Library unique name: <valkyrie.compose:codeviewer>
-final fun io.github.composegears.valkyrie.compose.codeviewer.core/CodeEditor(dev.snipme.highlights/Highlights, kotlin/Function1<kotlin/String, kotlin/Unit>, androidx.compose.ui/Modifier?, kotlin/Boolean, kotlin/Boolean, kotlin/Boolean, androidx.compose.ui.text/TextStyle?, kotlin/Function2<androidx.compose.runtime/Composer, kotlin/Int, kotlin/Unit>?, kotlin/Function2<androidx.compose.runtime/Composer, kotlin/Int, kotlin/Unit>?, kotlin/Function2<androidx.compose.runtime/Composer, kotlin/Int, kotlin/Unit>?, kotlin/Function2<androidx.compose.runtime/Composer, kotlin/Int, kotlin/Unit>?, kotlin/Boolean, androidx.compose.ui.text.input/VisualTransformation?, androidx.compose.foundation.text/KeyboardOptions?, androidx.compose.foundation.text/KeyboardActions?, kotlin/Boolean, kotlin/Int, kotlin/Int, androidx.compose.foundation.interaction/MutableInteractionSource?, androidx.compose.ui.graphics/Shape?, androidx.compose.material3/TextFieldColors?, androidx.compose.runtime/Composer?, kotlin/Int, kotlin/Int, kotlin/Int, kotlin/Int) // io.github.composegears.valkyrie.compose.codeviewer.core/CodeEditor|CodeEditor(dev.snipme.highlights.Highlights;kotlin.Function1<kotlin.String,kotlin.Unit>;androidx.compose.ui.Modifier?;kotlin.Boolean;kotlin.Boolean;kotlin.Boolean;androidx.compose.ui.text.TextStyle?;kotlin.Function2<androidx.compose.runtime.Composer,kotlin.Int,kotlin.Unit>?;kotlin.Function2<androidx.compose.runtime.Composer,kotlin.Int,kotlin.Unit>?;kotlin.Function2<androidx.compose.runtime.Composer,kotlin.Int,kotlin.Unit>?;kotlin.Function2<androidx.compose.runtime.Composer,kotlin.Int,kotlin.Unit>?;kotlin.Boolean;androidx.compose.ui.text.input.VisualTransformation?;androidx.compose.foundation.text.KeyboardOptions?;androidx.compose.foundation.text.KeyboardActions?;kotlin.Boolean;kotlin.Int;kotlin.Int;androidx.compose.foundation.interaction.MutableInteractionSource?;androidx.compose.ui.graphics.Shape?;androidx.compose.material3.TextFieldColors?;androidx.compose.runtime.Composer?;kotlin.Int;kotlin.Int;kotlin.Int;kotlin.Int){}[0]
-final fun io.github.composegears.valkyrie.compose.codeviewer.core/CodeViewer(dev.snipme.highlights/Highlights, androidx.compose.ui/Modifier?, androidx.compose.runtime/Composer?, kotlin/Int, kotlin/Int) // io.github.composegears.valkyrie.compose.codeviewer.core/CodeViewer|CodeViewer(dev.snipme.highlights.Highlights;androidx.compose.ui.Modifier?;androidx.compose.runtime.Composer?;kotlin.Int;kotlin.Int){}[0]
-final fun io.github.composegears.valkyrie.compose.codeviewer.core/getEmphasisLocations(kotlin/String, kotlin/String): kotlin/Array<dev.snipme.highlights.model/PhraseLocation> // io.github.composegears.valkyrie.compose.codeviewer.core/getEmphasisLocations|getEmphasisLocations(kotlin.String;kotlin.String){}[0]
-final fun io.github.composegears.valkyrie.compose.codeviewer.core/rememberCodeHighlight(kotlin/String, kotlin/Array<out dev.snipme.highlights.model/PhraseLocation>..., androidx.compose.runtime/Composer?, kotlin/Int): dev.snipme.highlights/Highlights // io.github.composegears.valkyrie.compose.codeviewer.core/rememberCodeHighlight|rememberCodeHighlight(kotlin.String;kotlin.Array<out|dev.snipme.highlights.model.PhraseLocation>...;androidx.compose.runtime.Composer?;kotlin.Int){}[0]
-final fun io.github.composegears.valkyrie.compose.codeviewer/KotlinCodeViewer(kotlin/String, androidx.compose.ui/Modifier?, kotlin/Function1<kotlin/String, kotlin/Unit>?, androidx.compose.runtime/Composer?, kotlin/Int, kotlin/Int) // io.github.composegears.valkyrie.compose.codeviewer/KotlinCodeViewer|KotlinCodeViewer(kotlin.String;androidx.compose.ui.Modifier?;kotlin.Function1<kotlin.String,kotlin.Unit>?;androidx.compose.runtime.Composer?;kotlin.Int;kotlin.Int){}[0]
