+++ sdk/compose/codeviewer/build.gradle.kts
@@ -8,11 +8,11 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(projects.sdk.compose.foundation)
+            implementation(projects.sdk.compose.highlightsCore)
             implementation(projects.compose.ui)
             implementation(projects.compose.util)
 
             implementation(compose.material3)
-            api(libs.highlights)
         }
     }
 }
