+++ sdk/compose/highlights-core/api/highlights-core.api
@@ -1,0 +1,5 @@
+public final class io/github/composegears/valkyrie/sdk/compose/highlights/core/CommonKt {
+	public static final fun buildAnnotatedString (Ldev/snipme/highlights/Highlights;)Landroidx/compose/ui/text/AnnotatedString;
+	public static final fun getEmphasisLocations (Ljava/lang/String;Ljava/lang/String;)[Ldev/snipme/highlights/model/PhraseLocation;
+}
+
