+++ sdk/compose/highlights-core/api/highlights-core.klib.api
@@ -1,0 +1,10 @@
+// Klib ABI Dump
+// Targets: [wasmJs]
+// Rendering settings:
+// - Signature version: 2
+// - Show manifest properties: true
+// - Show declarations: true
+
+// Library unique name: <valkyrie.sdk.compose:highlights-core>
+final fun (dev.snipme.highlights/Highlights).io.github.composegears.valkyrie.sdk.compose.highlights.core/buildAnnotatedString(): androidx.compose.ui.text/AnnotatedString // io.github.composegears.valkyrie.sdk.compose.highlights.core/buildAnnotatedString|buildAnnotatedString@dev.snipme.highlights.Highlights(){}[0]
+final fun io.github.composegears.valkyrie.sdk.compose.highlights.core/getEmphasisLocations(kotlin/String, kotlin/String): kotlin/Array<dev.snipme.highlights.model/PhraseLocation> // io.github.composegears.valkyrie.sdk.compose.highlights.core/getEmphasisLocations|getEmphasisLocations(kotlin.String;kotlin.String){}[0]
