+++ sdk/compose/highlights-core/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/compose/highlights/core/Common.kt
@@ -1,8 +1,5 @@
-package io.github.composegears.valkyrie.compose.codeviewer.core
+package io.github.composegears.valkyrie.sdk.compose.highlights.core
 
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
@@ -12,11 +9,8 @@ import dev.snipme.highlights.Highlights
 import dev.snipme.highlights.model.BoldHighlight
 import dev.snipme.highlights.model.ColorHighlight
 import dev.snipme.highlights.model.PhraseLocation
-import dev.snipme.highlights.model.SyntaxLanguage
-import dev.snipme.highlights.model.SyntaxThemes
-import io.github.composegears.valkyrie.compose.util.isLight
 
-internal fun Highlights.buildAnnotatedString() = buildAnnotatedString {
+fun Highlights.buildAnnotatedString() = buildAnnotatedString {
     append(getCode())
 
     getHighlights()
@@ -45,23 +39,6 @@ internal fun Highlights.buildAnnotatedString() = buildAnnotatedString {
                 end = it.location.end,
             )
         }
-}
-
-@Composable
-fun rememberCodeHighlight(
-    codeBlock: String,
-    vararg emphasisLocation: PhraseLocation,
-): Highlights {
-    val isLight = MaterialTheme.colorScheme.isLight
-
-    return remember(isLight, codeBlock) {
-        Highlights.Builder()
-            .code(codeBlock)
-            .language(SyntaxLanguage.KOTLIN)
-            .theme(SyntaxThemes.darcula(darkMode = !isLight))
-            .emphasis(*emphasisLocation)
-            .build()
-    }
 }
 
 fun getEmphasisLocations(
