+++ settings.gradle.kts
@@ -76,6 +76,8 @@ include("components:parser:kmp:xml")
 include("components:parser:kmp:svg")
 include("components:parser:unified")
 
+include("sdk:compose:codeviewer")
+include("sdk:compose:highlights-core")
 include("sdk:compose:foundation")
 include("sdk:core:extensions")
 include("sdk:core:xml")
@@ -90,7 +92,6 @@ include("sdk:ir:xml")
 include("sdk:test:coverage")
 include("sdk:test:resource-loader")
 
-include("compose:codeviewer")
 include("compose:icons")
 include("compose:ui")
 include("compose:util")
