+++ tools/idea-plugin/build.gradle.kts
@@ -28,10 +28,11 @@ dependencies {
     implementation(projects.components.generator.iconpack)
     implementation(projects.components.generator.jvm.imagevector)
     implementation(projects.components.parser.unified)
-    implementation(projects.compose.codeviewer)
     implementation(projects.compose.icons)
     implementation(projects.compose.ui)
     implementation(projects.compose.util)
+    implementation(projects.sdk.compose.codeviewer)
+    implementation(projects.sdk.compose.highlightsCore)
     implementation(projects.sdk.compose.foundation)
     implementation(projects.sdk.core.extensions)
     implementation(projects.sdk.intellij.psi.iconpack)
