+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/settings/tabs/generator/ui/SelectableCard.kt
@@ -15,7 +15,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import dev.snipme.highlights.Highlights
-import io.github.composegears.valkyrie.compose.codeviewer.core.rememberCodeHighlight
+import io.github.composegears.valkyrie.compose.util.isLight
+import io.github.composegears.valkyrie.sdk.compose.codeviewer.rememberCodeHighlight
 import io.github.composegears.valkyrie.sdk.compose.foundation.layout.CenterVerticalRow
 import io.github.composegears.valkyrie.ui.foundation.highlights.CodeViewerTooltip
 import io.github.composegears.valkyrie.ui.foundation.theme.PreviewTheme
@@ -73,6 +74,7 @@ private fun SelectableCardPreview() = PreviewTheme {
                     .height(24.dp)
                     .build()
                 """.trimIndent(),
+                isLight = MaterialTheme.colorScheme.isLight,
             ),
             isSelected = true,
             onSelect = {},
@@ -89,6 +91,7 @@ private fun SelectableCardPreview() = PreviewTheme {
                         .build()
                 }
                 """.trimIndent(),
+                isLight = MaterialTheme.colorScheme.isLight,
             ),
             isSelected = false,
             onSelect = {},
